+++ Pod/Classes/Internal/SEGLocation.m
@@ -10,21 +10,27 @@
 #import "SEGAnalyticsUtils.h"
 #import <CoreLocation/CoreLocation.h>
 
-#define LOCATION_STRING_PROPERTY(NAME, PLACEMARK_PROPERTY)                                      \
-    -(NSString *)NAME                                                                           \
-    {                                                                                           \
-        __block NSString *result = nil;                                                         \
-        dispatch_sync(self.syncQueue, ^{ result = self.currentPlacemark.PLACEMARK_PROPERTY; }); \
-        return result;                                                                          \
+#define LOCATION_STRING_PROPERTY(NAME, PLACEMARK_PROPERTY)     \
+    -(NSString *)NAME                                          \
+    {                                                          \
+        __block NSString *result = nil;                        \
+        dispatch_sync(self.syncQueue, ^{                       \
+            result = self.currentPlacemark.PLACEMARK_PROPERTY; \
+        });                                                    \
+        return result;                                         \
     }
 
-#define LOCATION_NUMBER_PROPERTY(NAME, PLACEMARK_PROPERTY)                                         \
-    -(NSNumber *)NAME                                                                              \
-    {                                                                                              \
-        __block NSNumber *result = nil;                                                            \
-        dispatch_sync(self.syncQueue, ^{ result = @(self.currentPlacemark.PLACEMARK_PROPERTY); }); \
-        return result;                                                                             \
+#define LOCATION_NUMBER_PROPERTY(NAME, PLACEMARK_PROPERTY)        \
+    -(NSNumber *)NAME                                             \
+    {                                                             \
+        __block NSNumber *result = nil;                           \
+        dispatch_sync(self.syncQueue, ^{                          \
+            result = @(self.currentPlacemark.PLACEMARK_PROPERTY); \
+        });                                                       \
+        return result;                                            \
     }
+
+#define LOCATION_AGE 300.0 // 5 minutes
 
 
 @interface SEGLocation () <CLLocationManagerDelegate>
@@ -45,12 +51,14 @@
 
     if (self = [super init]) {
         self.geocoder = [[CLGeocoder alloc] init];
-        self.locationManager = [[CLLocationManager alloc] init];
-        self.locationManager.delegate = self;
-#if !(TARGET_OS_TV)
-        [self.locationManager startUpdatingLocation];
-#endif
         self.syncQueue = dispatch_queue_create("io.segment.location.syncQueue", NULL);
+        dispatch_async(dispatch_get_main_queue(), ^{
+            self.locationManager = [[CLLocationManager alloc] init];
+            self.locationManager.delegate = self;
+#if !(TARGET_OS_TV)
+            [self.locationManager startUpdatingLocation];
+#endif
+        });
     }
     return self;
 }
@@ -65,6 +73,20 @@ LOCATION_NUMBER_PROPERTY(longitude, location.coordinate.longitude);
 #if !(TARGET_OS_TV)
 LOCATION_NUMBER_PROPERTY(speed, location.speed);
 #endif
+
+- (void)startUpdatingLocation
+{
+    if (self.locationManager && self.currentPlacemark) {
+        CLLocation *location = self.currentPlacemark.location;
+        NSDate *eventDate = location.timestamp;
+        NSTimeInterval howRecent = [eventDate timeIntervalSinceNow];
+        if (fabs(howRecent) > LOCATION_AGE) {
+            dispatch_async(dispatch_get_main_queue(), ^{
+                [self.locationManager startUpdatingLocation];
+            });
+        }
+    }
+}
 
 - (BOOL)hasKnownLocation
 {
@@ -91,14 +113,26 @@ LOCATION_NUMBER_PROPERTY(speed, location.speed);
 {
     if (!locations.count) return;
 
-    __weak typeof(self) weakSelf = self;
-    [self.geocoder reverseGeocodeLocation:locations.firstObject
-                        completionHandler:^(NSArray *placemarks, NSError *error) {
-                            __strong typeof(weakSelf) strongSelf = weakSelf;
-                            dispatch_sync(strongSelf.syncQueue, ^{
-                                strongSelf.currentPlacemark = placemarks.firstObject;
-                            });
-                        }];
+    //https://developer.apple.com/library/ios/documentation/UserExperience/Conceptual/LocationAwarenessPG/CoreLocation/CoreLocation.html
+    CLLocation *location = [locations lastObject];
+    NSDate *eventDate = location.timestamp;
+    NSTimeInterval howRecent = [eventDate timeIntervalSinceNow];
+    if (fabs(howRecent) < LOCATION_AGE) {
+        // If the event is recent, do something with it.
+        __weak typeof(self) weakSelf = self;
+        [self.geocoder reverseGeocodeLocation:location
+                            completionHandler:^(NSArray *placemarks, NSError *error) {
+                                if (error) {
+                                    SEGLog(@"error: %@", error);
+                                } else if (placemarks.count) {
+                                    __strong typeof(weakSelf) strongSelf = weakSelf;
+                                    dispatch_sync(strongSelf.syncQueue, ^{
+                                        strongSelf.currentPlacemark = placemarks.firstObject;
+                                        [strongSelf.locationManager stopUpdatingLocation];
+                                    });
+                                }
+                            }];
+    }
 }
 
 - (void)locationManager:(CLLocationManager *)manager didFailWithError:(NSError *)error